Tietojen tallennus ja lukeminen csv-tiedostoista

+++ app/src/main/java/com/example/sokerihiiri/ui/screens/settings/SettingsControlDataScreen.kt
@@ -14,9 +14,10 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.platform.LocalContext
+import androidx.documentfile.provider.DocumentFile
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
-import com.example.sokerihiiri.ui.components.RemoveDataDialog
+import com.example.sokerihiiri.ui.components.SeriousConfirmDialog
 import com.example.sokerihiiri.ui.components.SettingsBase
 import com.example.sokerihiiri.ui.navigation.Screens
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
@@ -33,6 +34,13 @@ fun SettingsControlDataScreen(
 ) {
     val context = LocalContext.current
     val settingsViewModel: SettingsViewModel = hiltViewModel()
+    val uiState = settingsViewModel.uiState
+
+    var showRemoveAllRoomDataDialog by remember { mutableStateOf(false) }
+    var showRemoveMeasurementsDialog by remember { mutableStateOf(false) }
+    var showRemoveInsulinInjectionsDialog by remember { mutableStateOf(false) }
+    var showRemoveMealsDialog by remember { mutableStateOf(false) }
+    var showLoadFileDialog by remember { mutableStateOf(false) }
 
     val writeCSVDirLauncher = rememberLauncherForActivityResult(
         contract = ActivityResultContracts.OpenDocumentTree())
@@ -43,17 +51,27 @@ fun SettingsControlDataScreen(
 
         }
     }
+    val readCSVLauncher = rememberLauncherForActivityResult(
+        contract = ActivityResultContracts.OpenDocument())
+    {  uri: Uri? ->
+        if (uri != null) {
+            settingsViewModel.setLoadingFileUri(uri)
+            val documentFile = DocumentFile.fromSingleUri(context, uri)
+            val fileName = documentFile?.name
+            if (fileName != null) {
+                settingsViewModel.setLoadingFileName(fileName)
+                settingsViewModel.checkFileContent(context, uri, snackbarHostState)
+                showLoadFileDialog = true
+            }
+        }
+    }
+
     val writePermissionState = rememberPermissionState(
         android.Manifest.permission.WRITE_EXTERNAL_STORAGE
     )
     val readPermissionState = rememberPermissionState(
         android.Manifest.permission.READ_EXTERNAL_STORAGE
     )
-
-    var showRemoveAllRoomDataDialog by remember { mutableStateOf(false) }
-    var showRemoveMeasurementsDialog by remember { mutableStateOf(false) }
-    var showRemoveInsulinInjectionsDialog by remember { mutableStateOf(false) }
-    var showRemoveMealsDialog by remember { mutableStateOf(false) }
 
     fun handledeleteAllRoomData() {
         settingsViewModel.deleteAllRoomData()
@@ -95,7 +113,22 @@ fun SettingsControlDataScreen(
     }
 
     fun handleLoadMeasurementsFromCSV() {
-
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+            readCSVLauncher.launch(arrayOf("*/*"))
+        } else {
+            // Kysyttävä lupaa jos api level 28 tai alle.
+            if (readPermissionState.status.isGranted) {
+                Log.d("SettingsControlDataScreen", "handleWriteCSV: Permission granted")
+                readCSVLauncher.launch(arrayOf("*/*"))
+            } else {
+                readPermissionState.launchPermissionRequest()
+                if (readPermissionState.status.shouldShowRationale) {
+                    Log.d("SettingsControlDataScreen", "handleWriteCSV: Permission denied, show rationale")
+                } else {
+                    Log.d("SettingsControlDataScreen", "handleWriteCSV: Permission denied")
+                }
+            }
+        }
     }
 
     SettingsBase(
@@ -105,7 +138,7 @@ fun SettingsControlDataScreen(
         Button(onClick = { handleWriteCSVButton() }) {
             Text(text = "Tallenna tiedot csv-tiedostoihin")
         }
-        Button(onClick = {handleLoadMeasurementsFromCSV()}) {
+        Button(onClick = { handleLoadMeasurementsFromCSV() }) {
             Text(text = "Lataa mittaustiedot tiedostosta")
         }
         Button(onClick = {
@@ -131,7 +164,7 @@ fun SettingsControlDataScreen(
     }
 
     if (showRemoveAllRoomDataDialog) {
-        RemoveDataDialog(
+        SeriousConfirmDialog(
             onConfirm = { handledeleteAllRoomData() },
             onDismiss = { showRemoveAllRoomDataDialog = false },
             title = "Poista kaikki tiedot",
@@ -140,7 +173,7 @@ fun SettingsControlDataScreen(
     }
 
     if (showRemoveMeasurementsDialog) {
-        RemoveDataDialog(
+        SeriousConfirmDialog(
             onConfirm = { handleDeleteAllMeasurementsConfirm() },
             onDismiss = { showRemoveMeasurementsDialog = false },
             title = "Poista kaikki mittaustiedot",
@@ -149,7 +182,7 @@ fun SettingsControlDataScreen(
     }
 
     if (showRemoveInsulinInjectionsDialog) {
-        RemoveDataDialog(
+        SeriousConfirmDialog(
             onConfirm = { handleDeleteAllInsulinInjectionsConfirm() },
             onDismiss = { showRemoveInsulinInjectionsDialog = false },
             title = "Poista kaikki insuliinikirjaukset",
@@ -158,11 +191,24 @@ fun SettingsControlDataScreen(
     }
 
     if (showRemoveMealsDialog) {
-        RemoveDataDialog(
+        SeriousConfirmDialog(
             onConfirm = { handleDeleteAllMealsConfirm() },
             onDismiss = { showRemoveMealsDialog = false },
             title = "Poista kaikki ateriatiedot",
             message = "Haluatko varmasti poistaa kaikki ateriatiedot? Tietoja ei voi palauttaa poistamisen jälkeen.",
             confirmText = "Ymmärrän ja haluan poistaa tiedot.")
+    }
+
+    if (showLoadFileDialog && uiState.loadingFileType != null) {
+        SeriousConfirmDialog(
+            onConfirm = {
+                showLoadFileDialog = false
+                settingsViewModel.readFileContent(context, snackbarHostState)
+            },
+            onDismiss = { showLoadFileDialog = false },
+            title = "Lataa tietoja csv-tiedostosta",
+            message = uiState.loadingFileReplaceWarning,
+            confirmText = "Kyllä"
+        )
     }
 }
